+++ main.js
@@ -1,15 +1,64 @@
-// noinspection JSUnusedGlobalSymbols
+/**
+ * @typedef {function} LoadScript
+ */
+/**
+ * @typedef {Object} Building
+ * @property {number} price
+ * @property {DocumentFragment} l
+ */
+/**
+ * @function l
+ * @param {string} id
+ */
+/**
+ * @function Beautify
+ * @param {number} number
+ * @param {number} decimalPlaces
+ */
+/**
+ * @function PlaySound
+ */
+/**
+ * @typedef {Object} Game
+ * @property {function} registerMod
+ * @property {function} Has
+ * @property {Object[]} GrandmaSynergies
+ * @property {Object} GrandmaSynergies.buildingTie
+ * @property {number} GrandmaSynergies.buildingTie.storedTotalCps
+ * @property {Object.<string, Building>} Objects
+ * @property {Building[]} ObjectsById
+ * @property {Array} customOptionsMenu
+ * @property {Array} Upgrades
+ * @property {string} clickStr
+ * @property {number} unbuffedCps
+ * @property {number} globalCpsMult
+ * @property {number} storedCps
+ */
+/**
+ * @typedef {Object} App
+ * @property {Array} mods
+ */
+/**
+ * @typedef {Object} CCSE
+ * @property {function} AppendCollapsibleOptionsMenu
+ */
+LoadScript(App.mods["Best Deal Helper"].dir + "/chroma.min.js");
 
 let BestDealHelper = {
     name: "BestDealHelper",
 
     config: {
-        sortbuildings: 0
+        sortbuildings: 0,
     },
 
     isLoaded: false,
+    load_chroma: false,
 
-    init: function () {
+    register: function () {
+        Game.registerMod(this.name, this);
+    },
+
+    "init": function () {
         MOD = this;
         Game.customOptionsMenu.push(MOD.addOptionsMenu);
         MOD.last_cps = 0;
@@ -22,34 +71,31 @@ let BestDealHelper = {
         MOD.isLoaded = true;
     },
 
-    load: function (str) {
+    "load": function (str) {
         const config = JSON.parse(str);
         for (const c in config) MOD.config[c] = config[c];
         MOD.sortBuildings();
     },
 
-    save: function () {
+    "save": function () {
         return JSON.stringify(MOD.config);
     },
 
-    register: function () {
-        Game.registerMod(this.name, this);
-    },
 
     addOptionsMenu: function () {
         const body = `
         <div class="listing">
-            ${MOD.button('sortbuildings', 'Sort Buildings ON (default)', 'Sort Buildings OFF')}
+            ${MOD.button("sortbuildings", "Sort Buildings ON (default)", "Sort Buildings OFF")}
         </div>`;
 
-        CCSE.AppendCollapsibleOptionsMenu(MOD.name, body)
+        CCSE.AppendCollapsibleOptionsMenu(MOD.name, body);
     },
 
     logicLoop: function () {
         if (
             MOD.last_cps !== Game.unbuffedCps
             || MOD.config.sortbuildings !== MOD.last_config_sortbuildings
-            || !document.querySelector('#normalizedCpspb0')
+            || !document.querySelector("#normalizedCpspb0")
         ) {
             MOD.sortBuildings();
             MOD.last_config_sortbuildings = MOD.config.sortbuildings;
@@ -60,7 +106,7 @@ let BestDealHelper = {
         let boost;
         let other;
         let synergyBoost = 0;
-        if (me.name === 'Grandma') {
+        if (me.name === "Grandma") {
             for (const i in Game.GrandmaSynergies) {
                 if (Game.Has(Game.GrandmaSynergies[i])) {
                     other = Game.Upgrades[Game.GrandmaSynergies[i]].buildingTie;
@@ -69,21 +115,35 @@ let BestDealHelper = {
                     synergyBoost += boost;
                 }
             }
-        } else if (me.name === 'Portal' && Game.Has('Elder Pact')) {
-            other = Game.Objects['Grandma'];
+        } else if (me.name === "Portal" && Game.Has("Elder Pact")) {
+            other = Game.Objects["Grandma"];
             boost = (me.amount * 0.05 * other.amount) * Game.globalCpsMult;
             synergyBoost += boost;
         }
         return me.storedCps * Game.globalCpsMult + synergyBoost / Math.max(me.amount, 1);
     },
+    median: function (values) {
+        if (values.length === 0) return 0;
 
+        values.sort(function (a, b) {
+            return a - b;
+        });
+
+        const half = Math.floor(values.length / 2);
+
+        if (values.length % 2) {
+            return values[half];
+        }
+
+        return (values[half - 1] + values[half]) / 2.0;
+    },
     insertAfter: function (newNode, referenceNode) {
         referenceNode.parentNode.insertBefore(newNode, referenceNode.nextSibling);
     },
 
     sortBuildings: function () {
-        let buildings = [...Game.ObjectsById]
-        buildings.forEach(e => e.cpsPerCookie = MOD.boosted(e) / e.price)
+        let buildings = [...Game.ObjectsById];
+        buildings.forEach(e => e.cpsPerCookie = MOD.boosted(e) / e.price);
 
         // Sort buildings or leave them to default
         if (MOD.config.sortbuildings) {
@@ -94,50 +154,62 @@ let BestDealHelper = {
         }
 
         // Sort buildings only if the order has changed
-        let buildings_order = buildings.map(e => e.id)
+        let buildings_order = buildings.map(e => e.id);
         if (!buildings_order.every((value, index) => value === MOD.last_buildings_order[index])) {
-            let store = document.querySelector('#products')
+            let store = document.querySelector("#products");
             for (let i = 0; i < buildings.length; ++i) {
                 store.appendChild(buildings[i].l);
             }
-            MOD.last_buildings_order = buildings_order
+            MOD.last_buildings_order = buildings_order;
             // Game.Notify(`Buildings are sorted!`, ``, [16, 5], 1.5, 1);
         }
 
         // Normalization by Mean
-        buildings = buildings.filter(o => o.locked === 0)
-        const avg = buildings.map(e => e.cpsPerCookie).reduce((a, b) => a + b, 0) / buildings.length;
-        buildings.forEach(e => e.cpsPerCookieDelta = e.cpsPerCookie / avg)
+        buildings = buildings.filter(o => o.locked === 0);
+        const cpsPerCookieArr = buildings.map(e => e.cpsPerCookie);
+        const avg = cpsPerCookieArr.reduce((a, b) => a + b, 0) / buildings.length;
+        let color = (
+            chroma.scale(["red", "yellow", "lightgreen"])
+                .mode("lrgb")
+                .domain([Math.min(...cpsPerCookieArr), MOD.median(cpsPerCookieArr), Math.max(...cpsPerCookieArr)])
+        );
+
         for (const i in buildings) {
             let me = buildings[i];
-            let cpspb = document.querySelector("#normalizedCpspb" + me.id)
+            let cpspb = document.querySelector("#normalizedCpspb" + me.id);
             if (!cpspb) {
                 cpspb = document.createElement("span");
                 cpspb.setAttribute("id", "normalizedCpspb" + me.id);
-                MOD.insertAfter(cpspb, l('productPrice' + me.id))
+                cpspb.style.fontWeight = "bolder";
+                MOD.insertAfter(cpspb, l("productPrice" + me.id));
             }
-            cpspb.textContent = "(💹" + Beautify(me.cpsPerCookieDelta * 100, 1) + "%)"
+            cpspb.textContent = "(💹" + Beautify(me.cpsPerCookie * 100 / avg, 1) + "%)";
+            cpspb.style.color = color(me.cpsPerCookie);
+
         }
     },
 
     button: function (config, texton, textoff) {
         const name = `BestDealHelper${config}button`;
-        const callback = `BestDealHelper.buttonCallback('${config}', '${name}', '${texton}', '${textoff}');`
+        const callback = `BestDealHelper.buttonCallback('${config}', '${name}', '${texton}', '${textoff}');`;
         const value = MOD.config[config];
-        return `<a class="${value ? 'option' : 'option off'}" id="${name}" ${Game.clickStr}="${callback}">${value ? texton : textoff}</a>`
+        return `<a class="${value ? "option" : "option off"}" id="${name}" ${Game.clickStr}="${callback}">${value ? texton : textoff}</a>`;
     },
 
-    buttonCallback: function (config, button, texton, textoff) {
+    "buttonCallback": function (config, button, texton, textoff) {
         const value = !MOD.config[config];
         MOD.config[config] = value;
-        l(button).innerHTML = value ? texton : textoff
-        l(button).className = value ? 'option' : 'option off'
-        PlaySound('snd/tick.mp3');
+        l(button).innerHTML = value ? texton : textoff;
+        l(button).className = value ? "option" : "option off";
+        PlaySound("snd/tick.mp3");
     },
 };
 
 // Bind methods
-for (func of Object.getOwnPropertyNames(BestDealHelper).filter(m => typeof BestDealHelper[m] === 'function')) {
+for (func of Object.getOwnPropertyNames(BestDealHelper).filter(m => typeof BestDealHelper[m] === "function")) {
+    /**
+     * @typedef {string} func
+     */
     BestDealHelper[func] = BestDealHelper[func].bind(BestDealHelper);
 }
 
@@ -146,7 +218,9 @@ if (!BestDealHelper.isLoaded) {
     if (CCSE && CCSE.isLoaded) {
         BestDealHelper.register();
     } else {
-        if (!CCSE) var CCSE = {};
+        if (!CCSE) { // noinspection JSUnusedLocalSymbols
+            let CCSE = {};
+        }
         if (!CCSE.postLoadHooks) CCSE.postLoadHooks = [];
         CCSE.postLoadHooks.push(BestDealHelper.register);
     }