+++ CHANGELOG.md
@@ -5,8 +5,13 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+
+## [2042.5] - 2021-09-15
+### Added
+- Colors!
+
 ### Changed
-- Sort buildings only if the order has changed
+- Efficiency: Sort buildings only if the order has changed
 
 ## [2042.4] - 2021-09-13
 ### Fixed